Storage/requirements.txt
connexion==2.14.1
swagger-ui-bundle==0.0.9
SQLAlchemy==1.4.42 
mysql-connector-python==8.0.23
pymysql==1.0.2
pykafka==2.8.0